LayoutTests/swipe/swipe-start-hysteresis-failures-expected.txt

swipe event (delta 0 0, phase 'maybegin')
scroll event (delta 1 0, phase 'began')
scroll event (delta 0 0, phase 'changed')
scroll event (delta 0 0, phase 'ended')
swipe event (delta 0 0, phase 'maybegin')
scroll event (delta 2 2, phase 'began')
scroll event (delta 0 0, phase 'changed')
scroll event (delta 0 0, phase 'ended')
swipe event (delta 0 0, phase 'maybegin')
scroll event (delta 0 2, phase 'began')
scroll event (delta 0 0, phase 'changed')
scroll event (delta 0 0, phase 'ended')
swipe event (delta 0 0, phase 'maybegin')
scroll event (delta 1 0, phase 'began')
scroll event (delta 0 2, phase 'changed')
scroll event (delta 0 0, phase 'ended')
swipe event (delta 0 0, phase 'maybegin')
scroll event (delta 3 1, phase 'began')
scroll event (delta 0 0, phase 'changed')
didBeginSwipe
scroll event (delta 0 0, phase 'ended')
swipe event (delta 0 0, phase 'ended')
